style: clippy 1.97 sweep
`make lint` had 60 findings across zeo-rt and zeo, invisible because
the clippy job never had clippy installed. Most are `cargo clippy --fix`
output. The rest by hand: type aliases for the complex statics, one
`allow(too_many_arguments)` per emitter entry, an unfulfilled `expect`
dropped, two orphaned doc blocks and one orphaned attribute removed, and
two redundant `let mut x = x` rebindings gone. No behavior change;
`make test-jit` is 7284/7284.

+++ crates/zeo-rt/src/builtins/array.rs
@@ -1651,7 +1651,6 @@ fn union_of(recv: &crate::collections::RArray, others: &[Vec<RubyValue>]) -> Vec
     out
 }
 
-
 /// A random index in `0..bound`, from the supplied RNG (`random.rand(bound)`)
 /// or the shared PRNG. `bound` is assumed nonzero by the callers.
 fn rand_below(random: &Option<RubyValue>, bound: usize) -> Result<usize, crate::Signal> {
@@ -1721,7 +1720,7 @@ pub fn array_aset_int_checked(
     } else {
         Err(index_error!(
             "index {i} too small for array; minimum: {}",
-            -(crate::array_len(arr) as i64)
+            -crate::array_len(arr)
         ))
     }
 }
